chore: Make gRPC a controller.
In a short sighted move, I made gRPC a snowflake instead of pushing into
the controller pattern already established. This lifts the gRPC server
into the controllers.
