+++ src/player.rs
@@ -26,26 +26,14 @@ use tokio::{
     sync::{oneshot, Mutex},
     task::JoinHandle,
 };
-use tonic::{transport::Server, Request, Response, Status};
 use tracing::{error, info, span, Level, Span};
 
+use crate::songs::Songs;
 use crate::{
     audio, config, dmx, midi,
     playlist::{self, Playlist},
     playsync::CancelHandle,
-    proto::player::v1::{
-        player_service_server::PlayerServiceServer, StatusRequest, StatusResponse,
-        FILE_DESCRIPTOR_SET,
-    },
     songs::Song,
-};
-use crate::{
-    proto::player::v1::{
-        player_service_server::PlayerService, NextRequest, NextResponse, PlayRequest, PlayResponse,
-        PreviousRequest, PreviousResponse, StopRequest, StopResponse, SwitchToPlaylistRequest,
-        SwitchToPlaylistResponse,
-    },
-    songs::Songs,
 };
 
 struct PlayHandles {
@@ -136,24 +124,6 @@ impl Player {
                     status_events,
                 ));
             }
-        }
-
-        let grpc = config.grpc();
-        if grpc.enabled() {
-            let addr = format!("0.0.0.0:{}", grpc.port()).parse()?;
-            let player = player.clone();
-            let reflection_service = tonic_reflection::server::Builder::configure()
-                .register_encoded_file_descriptor_set(FILE_DESCRIPTOR_SET)
-                .build_v1()?;
-            tokio::spawn(
-                Server::builder()
-                    .add_service(reflection_service)
-                    .add_service(PlayerServiceServer::new(player.clone()))
-                    .serve(addr),
-            );
-            info!("gRPC server started");
-        } else {
-            info!("gRPC disabled");
         }
 
         Ok(player)
@@ -536,9 +506,14 @@ impl Player {
     }
 
     /// Gets the all songs playlist used by the player.
-    #[cfg(test)]
     pub fn get_all_songs_playlist(&self) -> Arc<Playlist> {
         self.all_songs.clone()
+    }
+
+    /// Gets the play start time.
+    pub async fn get_play_start_time(&self) -> Option<SystemTime> {
+        let play_start_time = self.play_start_time.lock().await;
+        *play_start_time
     }
 
     /// Goes to the previous song and emits the MIDI event associated if one exists.
@@ -595,114 +570,6 @@ impl StatusEvents {
             }),
             None => None,
         })
-    }
-}
-
-// Playlist name constants.
-const PLAYLIST_NAME: &str = "playlist";
-const ALL_SONGS_NAME: &str = "all_songs";
-
-#[tonic::async_trait]
-impl PlayerService for Player {
-    async fn play(&self, _: Request<PlayRequest>) -> Result<Response<PlayResponse>, Status> {
-        match Self::play(self).await {
-            Some(song) => Ok(Response::new(PlayResponse {
-                song: Some(song.to_proto()?),
-            })),
-            None => Err(Status::failed_precondition("song already playing")),
-        }
-    }
-
-    async fn previous(
-        &self,
-        _: Request<PreviousRequest>,
-    ) -> Result<Response<PreviousResponse>, Status> {
-        let current_song = self.playlist.current();
-        let previous_song = Self::prev(self).await;
-
-        if current_song.name() == previous_song.name() {
-            return Err(Status::failed_precondition(
-                "can't move to previous song while playing",
-            ));
-        }
-
-        Ok(Response::new(PreviousResponse {
-            song: Some(previous_song.to_proto()?),
-        }))
-    }
-
-    async fn next(&self, _: Request<NextRequest>) -> Result<Response<NextResponse>, Status> {
-        let current_song = self.playlist.current();
-        let next_song = Self::next(self).await;
-
-        if current_song.name() == next_song.name() {
-            return Err(Status::failed_precondition(
-                "can't move to next song while playing",
-            ));
-        }
-
-        Ok(Response::new(NextResponse {
-            song: Some(next_song.to_proto()?),
-        }))
-    }
-
-    async fn stop(&self, _: Request<StopRequest>) -> Result<Response<StopResponse>, Status> {
-        match Self::stop(self).await {
-            Some(song) => Ok(Response::new(StopResponse {
-                song: Some(song.to_proto()?),
-            })),
-            None => Err(Status::failed_precondition("song not playing")),
-        }
-    }
-
-    async fn switch_to_playlist(
-        &self,
-        request: Request<SwitchToPlaylistRequest>,
-    ) -> Result<Response<SwitchToPlaylistResponse>, Status> {
-        let playlist_name = request.into_inner().playlist_name;
-        if playlist_name == PLAYLIST_NAME {
-            self.switch_to_playlist().await;
-            return Ok(Response::new(SwitchToPlaylistResponse {}));
-        }
-        if playlist_name == ALL_SONGS_NAME {
-            self.switch_to_all_songs().await;
-            return Ok(Response::new(SwitchToPlaylistResponse {}));
-        }
-
-        Err(Status::unimplemented(format!(
-            "only {} and {} are supported for now",
-            ALL_SONGS_NAME, PLAYLIST_NAME
-        )))
-    }
-
-    async fn status(&self, _: Request<StatusRequest>) -> Result<Response<StatusResponse>, Status> {
-        let playlist = self.get_playlist();
-        let playlist_name = if Arc::ptr_eq(&playlist, &self.playlist) {
-            PLAYLIST_NAME
-        } else {
-            ALL_SONGS_NAME
-        };
-
-        let elapsed = {
-            let play_start_time = self.play_start_time.lock().await;
-            match *play_start_time {
-                Some(play_start_time) => match play_start_time.elapsed() {
-                    Ok(play_start_time) => match prost_types::Duration::try_from(play_start_time) {
-                        Ok(play_start_time) => Some(play_start_time),
-                        Err(e) => return Err(Status::from_error(Box::new(e))),
-                    },
-                    Err(e) => return Err(Status::from_error(Box::new(e))),
-                },
-                None => None,
-            }
-        };
-
-        Ok(Response::new(StatusResponse {
-            playlist_name: playlist_name.to_string(),
-            current_song: Some(playlist.current().to_proto()?),
-            playing: elapsed.is_some(),
-            elapsed,
-        }))
     }
 }
